+++ tests/test_pipelines.py
@@ -54,10 +54,10 @@ def test_Pipeline__load_raw_gmr_data__sets_data(
     loaded_generic_pipeline, raw_generic_well_a1, raw_generic_well_a2
 ):
     np.testing.assert_array_equal(
-        loaded_generic_pipeline.get_raw_tissue_gmr(), raw_generic_well_a1
+        loaded_generic_pipeline.get_raw_tissue_magnetic_data(), raw_generic_well_a1
     )
     np.testing.assert_array_equal(
-        loaded_generic_pipeline.get_raw_reference_gmr(), raw_generic_well_a2
+        loaded_generic_pipeline.get_raw_reference_magnetic_data(), raw_generic_well_a2
     )
 
 
@@ -65,7 +65,7 @@ def test_Pipeline__load_raw_gmr_data__raises_error_if_data_already_loaded(
     loaded_generic_pipeline, raw_generic_well_a1, raw_generic_well_a2
 ):
     with pytest.raises(DataAlreadyLoadedInPipelineError):
-        loaded_generic_pipeline.load_raw_gmr_data(
+        loaded_generic_pipeline.load_raw_magnetic_data(
             raw_generic_well_a1, raw_generic_well_a2
         )
 
@@ -147,22 +147,22 @@ def test_Pipeline__get_sensitivity_calibrated_reference_gmr__calls_correct_metho
         (
             "abcde",
             "apply_noise_filtering",
-            lambda the_pipeline: the_pipeline.get_noise_filtered_gmr(),
+            lambda the_pipeline: the_pipeline.get_noise_filtered_magnetic_data(),
             [lambda the_pipeline: the_pipeline.get_fully_calibrated_gmr()],
             "applying noise filtering",
         ),
         (
             "qed",
             "compress_filtered_gmr",
-            lambda the_pipeline: the_pipeline.get_compressed_gmr(),
-            [lambda the_pipeline: the_pipeline.get_noise_filtered_gmr()],
+            lambda the_pipeline: the_pipeline.get_compressed_magnetic_data(),
+            [lambda the_pipeline: the_pipeline.get_noise_filtered_magnetic_data()],
             "applying compression",
         ),
         (
             "wakka",
             "calculate_voltage_from_gmr",
             lambda the_pipeline: the_pipeline.get_compressed_voltage(),
-            [lambda the_pipeline: the_pipeline.get_compressed_gmr()],
+            [lambda the_pipeline: the_pipeline.get_compressed_magnetic_data()],
             "converting compressed GMR data to voltage",
         ),
         (
@@ -176,7 +176,7 @@ def test_Pipeline__get_sensitivity_calibrated_reference_gmr__calls_correct_metho
             "trampoline",
             "peak_detector",
             lambda the_pipeline: the_pipeline.get_peak_detection_results(),
-            [lambda the_pipeline: the_pipeline.get_noise_filtered_gmr()],
+            [lambda the_pipeline: the_pipeline.get_noise_filtered_magnetic_data()],
             "detecting the peaks in the magnetic traces",
         ),
         (
@@ -185,7 +185,7 @@ def test_Pipeline__get_sensitivity_calibrated_reference_gmr__calls_correct_metho
             lambda the_pipeline: the_pipeline.get_magnetic_data_metrics(),
             [
                 lambda the_pipeline: the_pipeline.get_peak_detection_results(),
-                lambda the_pipeline: the_pipeline.get_noise_filtered_gmr(),
+                lambda the_pipeline: the_pipeline.get_noise_filtered_magnetic_data(),
             ],
             "calculate data metrics on the traces from the raw magentic readings",
         ),
@@ -240,7 +240,7 @@ def test_Pipeline__get_noise_filtered_gmr__creates_and_uses_filter_supplied_by_t
     mocked_function_under_test = mocker.patch.object(
         pipelines, "apply_noise_filtering", autospec=True,
     )
-    loaded_generic_pipeline.get_noise_filtered_gmr()
+    loaded_generic_pipeline.get_noise_filtered_magnetic_data()
     # Eli (7/6/20): NumPy arrays don't play well with assert_called_once_with, so asserting things separately
     assert mocked_function_under_test.call_count == 1
     actual_filter_array = mocked_function_under_test.call_args_list[0][0][1]
@@ -253,9 +253,9 @@ def test_Pipeline__get_noise_filtered_gmr__returns_same_data_if_no_filter_define
 ):
     no_filter_pipeline_template = PipelineTemplate(tissue_sampling_period=1000)
     pipeline = no_filter_pipeline_template.create_pipeline()
-    pipeline.load_raw_gmr_data(raw_generic_well_a1, raw_generic_well_a2)
+    pipeline.load_raw_magnetic_data(raw_generic_well_a1, raw_generic_well_a2)
     calibrated_data = pipeline.get_fully_calibrated_gmr()
-    filtered_data = pipeline.get_noise_filtered_gmr()
+    filtered_data = pipeline.get_noise_filtered_magnetic_data()
     assert filtered_data is calibrated_data